requirements.txt
gspread~=6.2.0
oauth2client~=4.1.3
google-api-python-client~=2.166.0
docxtpl~=0.19.1
docx2pdf~=0.1.8